Transfer receiving token from client to CommunicationThread
